chore: move pg differ under parser/differ/pg in parallel to mysql differ
